+++ spec/rails/templates/helpers_system_spec/app/controllers/home_controller.rb
@@ -1,0 +1,9 @@
+class HomeController < ApplicationController
+  def index
+    # nothing here
+  end
+
+  def foo
+    # nothing here
+  end
+end